Initial commit.
This should provide feature parity to the existing rabbithole implementation.
While not tested yet, it should also be compatible with exisiting DB schema.
Up-script (which is tested) was extracted from production database.

There is no implementation of migration yet (no need until there is an actual change to the database to migrate)
I have an idea for an improvement, but as none expects it for now, I will hold off in it, since it requires just that, a schema update.

The improvement is to store message headers as well. Don't think anything we do uses them, but maybe we should.